profile home functional desgin phase one

+++ pages/api/addCrypto.js
@@ -7,14 +7,13 @@ export default async function handler(req, res) {
 
     const new_stock = req.body
     console.log(new_stock.ticker)
-    
+
     async function main() {
         const userstock = await prisma.Usercrypto.create({
-            data:
-            {
-                name:new_stock.name,
-                shares:parseInt(new_stock.shares),
-                owner:{connect:{username: new_stock.username}}
+            data: {
+                name: new_stock.name,
+                shares: parseFloat(new_stock.shares),
+                owner: { connect: { username: new_stock.username } }
             }
         })
     }
